+++ nym-vpn-app/src-tauri/src/fs/mod.rs
@@ -1,12 +1,5 @@
 pub mod app;
-// #[cfg(windows)]
-// pub mod apps_windows;
-
-// pub mod app_discovery;
-
+pub mod app_discovery;
 pub mod config;
 pub mod path;
-// pub mod tunnel;
 pub mod util;
-
-pub mod app_discoveryold;
